requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-google-genai>=2.0.0
faiss-cpu>=1.7.4
google-generativeai>=0.3.0
pypdf>=3.17.0
python-docx>=1.1.0
tiktoken>=0.5.0
python-dotenv>=1.0.0
nest-asyncio>=1.5.8

